cleanup(sidekick): consistent management of WKT (#2258)
All the current service specifications (Protobuf, OpenAPI, Discovery)
use the same WKT (well-known types). In Protobuf the proto descriptors
of the WKT are loaded by the `import "google/protobuf/...";` directives.
In other specifications, the types need to be defined by the parser.
Likewise, in tests we need to define these types ahead of time.