feat(sidekick): enum serialization for non-proto sources (#2410)
Change the serialization of enums for non-proto sources (discovery docs
and OpenAPI) to use strings. The main motivation to use integers in the
case of Protobuf is roundstrips where we receive an unknown enum from
a Protobuf / gRPC response, copy that to a request, and want to
serialize it without loss. That scenario is not a consideration for
OpenAPI or Discovery docs, and serializing enums as integers just does
not work.