Revert "feat(sidekick): Inject InstrumentationClientInfo for tracing" (#2435)
Reverts googleapis/librarian#2252

My automated testing was wrong, it was happily disabled in both the
lib.rs and transport.rs because the feature flag wasn't propagated. I
have prepared a fix for the generator, but I'm finding this code
incorrect and so I'd like to revert if it can be done cleanly.