fix: Conventional Commit parser filters files based on Librarian flow (#2433)
Fixes: https://github.com/googleapis/librarian/issues/2403

Generation PR is based on files changed in the source repo (googleapis)
and NOT language repo. This should filter for files changed that match
the the paths for each API in a library.

Release PR is based on files changed in the language repo and NOT source
repo. This should filter for files changed that match the the paths for
a library's `source_root`.

## Changes
- Introduce a parameter for generation / release flows to define their
own files filtering logic (apiPaths vs sourceRoot)

---------

Signed-off-by: Lawrence Qiu <lawrenceqiu@google.com>
Co-authored-by: Joe Wang <106995533+JoeWang1127@users.noreply.github.com>