fix(internal/conventionalcommits): keep first value for repeated footer keys (#2440)
Footer with repeated keys is not supported with current implementation,
this fix let the first value be kept instead of last. Clarify in doc
comment about this limitation.

When onboarding libraries that has generation pr with owlbot and release
with Librarian,
[commit](https://github.com/googleapis/google-cloud-python/commit/34a7916fa58727a01a7d32c6a1adff59a351fd54),
the first `Source-Link:` is useful.

Alternative is to refactor `Footer` with type `map[string][]string`.
It's probably not worth it because the keys we care for Librarian are
not expected to repeat. This can be added as enhancement when needed.

For #2080