fix(librarian): don't add commit to generate PR if no revevant file changed (#2522)
Added a new filter to the generate commit analyzer. We now check that
the most recent commit, the regeneration commit, contains files related
to a particular library. If it does not it will not show up as a line
items for that library in the generation PR.

Additionally:
- Updated some variable names to make it easier to distinguish between
source and language repos

Fixes: #2493