fix(sidekick): Fixes several issues with oneof main setter samples (#2589)
- Uses the right import for the oneof group
- Takes boxing of the oneof message variants into account.
- Messages variants are now deprioritized for selection as a oneof
example variant.
- Makes the values shown match the ones used in other samples

This also removes the sample generation for when the example one of
varinat that's use on the main setter is repeated or map, as that needs
bigger improvements that will be done in a subsequent PR. Note that the
samples for each specific variant, including repeated and map, continue
to be generated.