fix: omit status field when empty (#2654)
This PR omits that `status` field when it is empty. The `status` field
is always empty in `generate-request.json`, so this change ensures that
`generate-request.json` no longer contains the `status` field. This
aligns with what we have in the docs:
https://github.com/googleapis/librarian/blob/main/doc/language-onboarding.md

We only need the `status` field in `configure-request.json`, and the
value is expected to be either `new` or `existing` as per the above
docs.