feat(internal/container/java): generate pom.xml files (#2682)
Add creation of `pom.xml` files for the generated Java client libraries.
We assume that the `google-cloud-pom-parent` and
`google-cloud-jar-parent` are published from a separate repository. For
now, the version is hard-coded in the templates. The intent here is that
we will no longer have to distinguish between libraries in the monorepo
vs split repos in terms of parent poms.

`run-genrate-library.sh` is also updated to not only generate the sample
library, but to build it as well.

Fixes #2657