feat(librarian): write a timing log at the end of update-image (#2771)
This writes a file rather than logging the timing of each library to
avoid cluttering up the logs, and to make it easier to look at just
timing.

At the moment this only writes a timing log for update-image, but the
writeTiming function is designed to make it easy to call from any other
comamnd (e.g. release init or generate) if we want.