feat(internal/librarian): add test flag to update-image command (#2729)
Adds `test` flag to update-image command to trigger test flow after
generation before commit and push.

It creates a temporary commit to ensure a clean repo state for the test
runner, runs the tests, and then soft-resets the commit to leave the
working directory in its original dirty state for the final commit
operation. If test fails, the repo will be left with the temp changes
created by test, on top of the temp commit for update-image generation.

Will update onboarding doc separately.

Fix #2624