fix(internal/librarian): test logic should look for service and message/enums separately (#2809)
With this change, the test now tracks services separately from
messages/enums in proto files and provides more coverage.
Instead of only tracking a comment change for the first found
message/enum/service, it now tracks a first service and first
message/enum if both are present in the proto file.

Fix #2623