feat(librarian): allow update-image to be a no-op (#2807)
The update-image command is really useful for container authors, in
terms of checking performance, behavior, and configuration. It's useful
to be able to run it even without a new image - because if everything is
in a steady state, we shouldn't see any changes at all.

We still log that the image is the same, but don't count it as an error.