chore: rename e2etest tag to mock_github (#2827)
Renaming `e2etest` tag to `mock_github` for clarity.

I find it confusing that we have both `e2e` tag and `e2etest` tag. 
The `e2e` tag is used for end to end test and is a widely used Go
convention.
The `e2etest` tag was added in #2228 to switch between a mock
implementations of repository for e2e test to avoid real API calls and
the real implementation for production code.