test: add integration build tag (#2831)
This test verify the integration of the librarian tool with external
systems, GitHub and Artifact Registry. It has expectation setups in env
vars and auth to properly work.
With an `integration` build tag, all tests in system_test.go will be
skipped when running `go test ./...` locally. Which is same behavior for
most developers who does not have relevant env vars setup.

Also remove the temp skip for TestFindLatestImage because it is no
longer triggered with standard `go test ./...`. Added a note to make
debugging easier.

Fixes #2720