feat(surfer): create surfer generate command framework (#2834)
Create the initial surfer CLI and add the framework for `surfer
generate`. The generation logic will be added in a follow-up change.

Remove gcloud.Generate from sidekick, since we no longer intend to run
it from the sidekick CLI.

Move `testdata/gcloud.yaml` to `testdata/parallelstore/gcloud.yaml` to
make it clear that it belongs to the parallelstore service.

For https://github.com/googleapis/librarian/issues/2375