chore: update stage-release automation to run weekly instead of bi-weekly (#2836)
Originally we had set a bi-weekly cadence in anticipation of needing to
support java and its release train. Now that it seems that java will not
be using a release train, let's make the releases weekly to be more
timely and consistent.

Note this also removes the forceRun flag which was used in the case you
wanted automation to force run stage-release command on an even week.