refactor: move e2e_test.go and its testdata to internal (#2934)
Move `e2e_test.go` and related test data to internal/integration.
- Change package to `integration_test` following Go's convention for
external test packages. This ensures e2e test only interact with the
exported APIs (which it is already doing).
- The `e2e` specific test data has been moved from
`internal/testdata/e2e` to `internal/integration/testdata/e2e`. This
adheres to the principle of co-location, keeping test data alongside the
tests that use it.
- Update testing doc and workflow to reflect changes.

The rest of testdata should also be moved to out of root into `internal`
and co-locate in the package of the test, will do that in a separate PR.

For #2842