feat(internal/command): add command execution package with logging (#2968)
An internal/command package is created to execute external commands and
print them to stderr before execution, in order to provide progress
reports to the user.

Existing call sites from internal/sidekick/external are migrated to use
the new package and the old package is deleted.

For https://github.com/googleapis/librarian/issues/2966