refactor(internal/legacylibrarian): move code (#2977)
Move code for librarian into internal/legacylibrarian, so that we can
use the namespace for new librarian code.

The cmd/automation and cmd/librarian binaries will be renamed in a
follow up PR, so that we can coordinate changes in Dockerfiles if
needed.

For https://github.com/googleapis/librarian/issues/2975