feat(librarian): add generate command (#3025)
An initial version of the generate command is added to the librarian
CLI, which generates client libraries based on configuration in
librarian.yaml.

The command supports two modes:

    librarian generate <library-name>: generates a single library
    librarian generate --all: generates all libraries

A testhelper generator is provided in order to verify the generate
command logic without calling real language-specific tooling.

Support for rust.Generate is also implemented.

For https://github.com/googleapis/librarian/issues/2966