feat: add create command to librarian (#3229)
This introduces the placeholder create command with all input args, that
checks to see if a library already exists, if so, invoke generate
command, if not it will invoke rust specific create logic (not
implemented).

This also introduces an interface for the generate command that can be
mocked out for testing purposes.

For #3072