librarian: move functions to detect changes out of sidekick to be reused (#3264)
This change is part 2 of changes to port over rust-publish for Rust. 

Move functions to detect changes (based on git commands) to a new change
package, as preparation for implementation. Originals in:
[last_tag.go](https://github.com/googleapis/librarian/blob/main/internal/sidekick/rust_release/last_tag.go),
[changes.go](https://github.com/googleapis/librarian/blob/main/internal/sidekick/rust_release/changes.go).
Migrate sidekick to use these moved functions.



For https://github.com/googleapis/librarian/issues/3153

Draft because it has repeat changes to #3262. Will fix once that is
merged