impl(rust): option to extend gRPC transport stub (#3282)
Motivated by https://github.com/googleapis/google-cloud-rust/issues/3957

Extending an existing transport stub lets us use as much generated code
as possible, while still adding on a handwritten bit for the
bidirectional streaming RPC. We haven't yet figured out a
one-size-fits-all solution for bidi streams in GAPICs, so this unblocks
progress on Pub/Sub).

---

I confess to being confused by the `rust-migrate` stuff. I put this
config setting on a `RustModule` and not a `RustCrate`, because it is
not really applicable for GAPICs, only for veneers.