chore: add handling for converting newlines in documentoverrides from toml to yaml (#3284)
Newlines were getting stripped from documentoverrides field. Originally
I had tried the "proper" solution to set this field to LiteralStyle in
yaml which should then keep the newlines. However that was not working
properly when we read and wrote back to the file.. Since this is a 1
time migration script I decided just to go with a brute force solution
to unblock us as this code will not be used once we migrate the sidekick
repo.

Fix #3136