fix(devtools/cmd/migrate-sidekick): add hardcoded exclusion list for veneer detection (#3317)
Adds a hardcoded list to excluded when identifying veneer libraries. The
migrate tool treats any library with a Cargo.toml file outside the
src/generated directory as a veneer library. This heuristic approach is
a bit too broad and included libraries that are not generated:
-
[echo-server](https://github.com/googleapis/google-cloud-rust/blob/main/src/gax-internal/echo-server/src/lib.rs):
test util that is handwritten
-
[gcp-sdk](https://github.com/googleapis/google-cloud-rust/blob/main/src/root/Cargo.toml):
workspace Crate

For #3313