chore: refactor release cmd so its source of truth is librarian.yaml not cargo files (#3316)
"release -all" was failing because it was trying to find releases for
packages specified in cargo.toml file but then not finding them in
librarian.yaml. Cargo.toml files should not be source of truth as it
contains entries that don't map to a library, ex:
[user-guide-samples](https://github.com/googleapis/google-cloud-rust/blob/c37cdf6e9c80acd9f1501e1d39714d31e4f07ac4/Cargo.lock#L7939).

This refactor changes so that we use librarian.yaml as the source of
truth for getting libraries to release. It moves this logic to librarian
release.go file as this should not be language specific. As part of
moving this logic I also migrated all release-all logic to librarian
release.go, since this is also is not language specific.

Example PR after running release-all:
https://github.com/googleapis/google-cloud-rust/compare/main...ldetmer:google-cloud-rust:test-release?expand=1

Fixes #3308

---------

Signed-off-by: ldetmer <1771267+ldetmer@users.noreply.github.com>
Co-authored-by: Min Zhu <zhumin@google.com>