fix(devtools/cmd/migrate-sidekick): fix mismatch in newline handling for documentationoverrides (#3341)
Add on a temp fix on top of the previous fix #3284. This fix tries to
align the expected format for `documentation_overrides` for
`google-cloud-orgpolicy-v1`. This is a brittle change, but as we don't
expect new `documentation_overrides` in the next few days and this is a
one time migration, it should be acceptable.


Verified locally after this change, `librarian generate
google-cloud-orgpolicy-v1` produces no diff. And the only diff it
produced for librarian.yaml is the following:

```
--- a/librarian.yaml
+++ b/librarian.yaml
@@ -1084,14 +1084,16 @@ libraries:
       title_override: Organization Policy Types
       documentation_overrides:
         - id: .google.cloud.orgpolicy.v1.Policy.ListPolicy
-          match: 'Ancestry subtrees must be in one of the following formats:'
+          match: |
+            Ancestry subtrees must be in one of the following formats:
           replace: |
-            
+
             Ancestry subtrees must be in one of the following formats:
         - id: .google.cloud.orgpolicy.v1.Policy.ListPolicy
-          match: The `supports_under` field of the associated `Constraint`  defines whether
+          match: |
+            The `supports_under` field of the associated `Constraint`  defines whether
           replace: |
-            
+
             The `supports_under` field of the associated `Constraint`  defines whether
```

Additional context about this fix:
https://github.com/googleapis/librarian/issues/3330#issuecomment-3687570409

Fix #3330